Starting implementing code to get the MethodInfo for the method being intercepted.
I'll use this info (and probably the actual arguments) to pass to an interceptor for deciding whether or not the method should be intercepted
or the original method should be invoked. Right now a problem I seem to be having is that the 'GetMethod' method of System.Type cannot be
found by TypeHelper.