Fixed method resolution problem.
TypeHelper wasn't finding the method because the array type (VectorType) I was creating didn't have the intern factory correctly set. Now, when an array is a parameter,
TypeHelper will find the correct method. I left somet pretty useless code in to try to find the method by 'brute force', which never worked in the first place. It might
come in handy some time in the future, so I'm checking it in now and will delete it later.