Checking in broken starting implementation with if-branching in replacement method.
I'm checking this in to preserve what I did with the IfBranch stuff. Right now I think I'm going to avoid this logic. I'll just pass in the method info and parameter
type information to the should intercept, then keep with the original logic that assumed interception. Some main interceptor will implement the branching logic
so that the original method is called with the original arguments and 'invocation.Return' holds the result (where applicable).