Added preliminary expectation recording.
I found an error with the test that detected if a method was being intercepted in a test (in Faker.CallsTo). I fixed that. There was an error with how
the method was being rewritten for those interceptions. I fixed that. I added some implementation for RecordingInterceptor (not thread safe).