Added integration tests.
It's still hacky and preliminary, but I can now specify a method to intercept and replace it with another method. The method call (in the test) will get recorded,
and when the code under test is executed, the substitute method will be called. I'm still trying to figure out the best way to do all of the expectation
recording and the proper syntax. The commented out Faker method is cleaner on the backend and adopts the style I've seen in AutoMapper and other places, but
I have some reservations about it; I still prefer a series of chained methods.