Updated binaries and fixed critical error.
I discovered an 'ExecutionEngineException' (nasty!) resulting from declaring an improper variable type to be returned. The strange thing is, though it fails peverify,
the code will work for cil primitives (int, bool, string, etc). It won't work for other value types (decimal, datetime, etc). I never even checked to see if it would
work for reference types. I corrected the bug, and added the MscorblibDependencyTests, which will be expanded shortly.

During the course of trying to track down the bug with my code, I updated the CCI assemblies to latest and discovered that backward compatibility with CodeMutatingVisitor
was broken. I reported this as a bug on the CCI codeplex site, and Herman fixed it the same day. All binaries are currently up to date and working. On his recommendation,
however, I need to look at using a different code walking mechanism to do my transformations, and at the very least, the scanning. I took a look at BaseCodeTraverser
(or something like that) for this task, and it should work well, at least for the scanning.