Added support for static properties.
This is really ugly. I had to provide an overload to 'CallsTo' that takes a generic function. Unfortunately, it's really hard to match that generic definition
with the actual call at runtime. Also, it adds a wrinkle in that I have to handle an extra case when examining the method body. Blah blah... but it seems to work
for now. File it under 'refactor later'.