Got sealed types working.
The interception for methods on sealed types should work for *any* instance methods, but I don't currently want to test that theory...
This check-in is very messy: checks for 'if IsStatic' all over the place need to be refactored, Faker needs to be cleaned up because
I originally thought my problem was with passing arguments to DynamicInvoke, and the common parts of the Registrars and Replacers need
to be consolidated somehow.

One major improvement that I need to make eventually is only intercepting the specified calls. This shouldn't be too hard... I'm not sure why
I haven't done it yet.