Cleaned up Faker a little bit.
I also wanted to add these notes: currently, some assemblies fail PE verification. I think that might be because I'm declaring a generic type with
no type parameters sometimes (when there is a void method with no parameters). The implementation still works, but peverify flags the assembly with
metadata errors (I may be way off about the reason, but this definitely happens when sealed types are involved). Anyway, I need to add a peverify
step to the tests!