Most important: I finally fixed the 'arity' issue. My variable declaration for VoidAction was wrong, it was a generic type without parameters.
I added a bunch of stuff to try to help me debug. The CodePrinter is no good right now.

Next I need to figure out what the difference is between the new assembly construction classes
and the existing ReplacementMethodBuilder code. Eventually I want to use the assembly construction
classes in the method builders.